+++ skia-safe/src/gpu/ganesh/types.rs
@@ -1,48 +1,21 @@
-use skia_bindings as sb;
 use std::ptr;
 
-// TODO: this should be a newtype(bool) I guess with implementations
-//       of From<bool> and Deref?
-pub use skia_bindings::skgpu_Protected as Protected;
-variant_name!(Protected::Yes);
+use crate::gpu;
+use skia_bindings as sb;
 
-// TODO: this should be a newtype(bool) I guess with implementations
-//       of From<bool> and Deref?
-pub use skia_bindings::skgpu_Renderable as Renderable;
-variant_name!(Renderable::No);
-
-pub use skia_bindings::skgpu_BackendApi as BackendApi;
-variant_name!(BackendApi::Metal);
-
-#[repr(C)]
-#[derive(Copy, Clone, PartialEq, Eq, Debug)]
-pub struct Budgeted(bool);
-
-native_transmutable!(sb::skgpu_Budgeted, Budgeted, budgeted_layout);
-
-#[allow(non_upper_case_globals)]
-impl Budgeted {
-    #[deprecated(since = "0.29.0", note = "use No")]
-    pub const NO: Budgeted = Budgeted(false);
-    #[deprecated(since = "0.29.0", note = "use Yes")]
-    pub const YES: Budgeted = Budgeted(true);
-
-    // we want this look like enum case names.
-    pub const No: Budgeted = Budgeted(false);
-    pub const Yes: Budgeted = Budgeted(true);
-}
-
-// TODO: CallbackResult
-
-// TODO: this should be a newtype(bool) I guess with implementations
-//       of From<bool> and Deref?
-pub use skia_bindings::skgpu_Mipmapped as Mipmapped;
-
-pub use skia_bindings::skgpu_Origin as Origin;
-variant_name!(Origin::TopLeft);
-
-pub use skia_bindings::GrBackendApi as BackendAPI;
+pub use skia_bindings::GrBackendApi as BackendApi;
 variant_name!(BackendAPI::OpenGL);
+
+#[deprecated(since = "0.0.0", note = "use BackendApi")]
+pub use BackendApi as BackendAPI;
+
+pub const METAL_BACKEND: BackendApi = BackendApi::Metal;
+pub const VULKAN_BACKEND: BackendApi = BackendApi::Vulkan;
+pub const MOCK_BACKEND: BackendApi = BackendApi::Mock;
+
+pub use gpu::Renderable;
+
+pub use gpu::Protected;
 
 pub use skia_bindings::GrSurfaceOrigin as SurfaceOrigin;
 variant_name!(SurfaceOrigin::BottomLeft);
@@ -90,7 +63,7 @@ pub use sb::GrMarkFrameBoundary as MarkFrameBoundary;
 variant_name!(MarkFrameBoundary::Yes);
 
 #[repr(C)]
-#[derive(Copy, Clone)]
+#[derive(Copy, Clone, Debug)]
 pub struct SubmitInfo {
     pub sync: SyncCpu,
     pub mark_boundary: MarkFrameBoundary,
